requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
cffi==1.17.1
cryptography==43.0.3
diff-match-patch==20241021
Django==4.2.16
django-environ==0.11.2
django-import-export==4.1.1
django-simple-history==3.7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pycparser==2.22
PyJWT==2.9.0
sqlparse==0.5.1
tablib==3.5.0
typing-extensions==4.12.2
